requirements.txt
# Core ML Libraries
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
datasets>=2.0.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Data Handling
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress Bars & Utilities
tqdm>=4.62.0
wandb>=0.12.0  # For experiment tracking (optional)

# Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0

# Configuration
PyYAML>=6.0
omegaconf>=2.1.0  # For advanced config management

# Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0

# Code Quality (optional)
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# API/Web (if you plan to deploy)
fastapi>=0.75.0
uvicorn>=0.17.0
